🐛 fix broken color table
